requirements.txt
APScheduler==3.10.4
playwright==1.48.0
psycopg2==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.8
